+++ docs/CHANGES.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 
 - `resources.default.path_maker` is now `pm_file` instead of `pm_class`. This only affects resources relative to a test `some_package.TestSomeThing.test_an_aspect` where the base path is now `some_package/TestSomeThing__test_an_aspect` instead of `some_package__TestSomeThing/test_an_aspect`. This gives fewer directories next to the test file and less clutter.
+- Made `utils._coalesce` and `utils._JSON_PREPPERS` private.
+- Made `resources._accept_count` private.
 
 ## [0.9.0] - 2025-10-21